Cache the entry points discovered within a namespace
Scanning the entry point registry is relatively expensive
and causes performance issues with unit tests of code
depending on stevedore. This change addresses the
performance issues by caching the entry points as
they are loaded from pkg_resources in a class attribute
in the base class of the extension managesr so they can
be reused by other instances.

Change-Id: Iba7bee6790cdedc94cb537e2ed6e12219c85f26a
Signed-off-by: Doug Hellmann <doug.hellmann@dreamhost.com>