Removed non-free color profile from .jpg
All the .jpg files in doc/source/essays had this embedded:
"Copyright 2007 Apple Inc., all rights reserved."

This is an ICC color profile which is copyrighted by Apple
and which is non-free, and may trigger some warnings/errors
by the distribution linters.

This patch converts the .jpg into the sRGB color space. I
did this simply loading the images in GIMP, and exporting
again after accepting the auto-coversion to sRGB.

Change-Id: Ibfdf0828ff149d5c83eb864c4fb854e80868a613