Use Stevedore exceptions for finding extensions
When there are no extensions or multiple extensions found for a
namespace stevedore raises RuntimeErrors. This means that consuming
applications must catch RuntimeErrors when loading plugins which may
mask errors coming from the plugin itself.

Instead have stevedore raise an error specific to the failure case that
can be handled from an application.

Change-Id: I0ab0556a589453bd8631d77f42506100b278a83d
Closes-Bug: #1501324