make error reporting for extension loading quieter
Only log the full traceback for an import error from an extension if
debug logging is enabled. This avoids spurious tracebacks in the log
when an application probes for all available extensions or when a
dependency of an extension is not installed (in that case the error
message itself should be sufficient to debug the issue).

Change-Id: I43c24f3791d575c2ecf0d08d82d9de5d74ec1150
Signed-off-by: Doug Hellmann <doug@doughellmann.com>