Allow suppression of warnings from DriverManager
Some projects (e.g. Neutron) using the DriverManager with
another fallback mechanism when the alias doesn't exist don't
want the "Could not load warning" messages in the logs.

This patch allows the warn_on_missing_entrypoint kwarg to be
passed to DriverManager which passes it up to its parent
NamedExtensionManager.

Change-Id: Ia6f5f749fc2f73ca6091fa6d58506fddb058902a