fix tox python3 overrides
We want to default to running all tox environments under python 3, so
set the basepython value in each environment.

We do not want to specify a minor version number, because we do not
want to have to update the file every time we upgrade python.

We do not want to set the override once in testenv, because that
breaks the more specific versions used in default environments like
py35 and py36.

Decode subprocess output so it has the right type under python 3.

Change-Id: Id2df5439fa6e788199bec4ac7c98c56347053621
Signed-off-by: Doug Hellmann <doug@doughellmann.com>